Add border and description text to drb container, rename to DRB
