+++ src/config/constants.ts
@@ -9,6 +9,8 @@ export const DRB_API_NAME = "drb"
 // URLs
 export const CIRCULATING_CATALOG_URL = "https://nypl.na2.iiivega.com/"
 export const LEGACY_CATALOG_URL = "https://legacycatalog.nypl.org/"
+export const DRB_ABOUT_URL =
+  "https://digital-research-books-beta.nypl.org/about?source=catalog"
 
 // API Routes
 export const DISCOVERY_API_SEARCH_ROUTE = "/discovery/resources"
