+++ styles/components/DRB.module.scss
@@ -1,0 +1,4 @@
+.drbContainer {
+  border-radius:  var(--nypl-space-xs);
+  padding: var(--nypl-space-s);
+}
